Move misplaced use case tests into subfolders
